pyproject.toml
[tool.poetry]
name = "personik-bot"
version = "0.1.0"
description = ""
authors = ["Max Bazarov <bazaroffm@gmail.com>"]
readme = "README.md"
packages = [{include = "personik_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
isort = "^5.12.0"
python-telegram-bot = "^20.4"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
